+++ AddOns/DBM-Sepulcher/LordsofDread.lua
@@ -1,7 +1,7 @@
 local mod	= DBM:NewMod(2457, "DBM-Sepulcher", nil, 1195)
 local L		= mod:GetLocalizedStrings()
 
-mod:SetRevision("20220330031553")
+mod:SetRevision("20220406015212")
 mod:SetCreatureID(181398, 181399)
 mod:SetEncounterID(2543)
 mod:SetUsedIcons(1, 2, 6, 7, 8)
@@ -51,7 +51,7 @@ local specWarnCloudofCarrion					= mod:NewSpecialWarningMoveAway(360012, nil, ni
 local specWarnCloudofCarrionDebuff				= mod:NewSpecialWarningYou(360012, nil, nil, nil, 1, 2)
 local specWarnCloudofCarrionDebuffMove			= mod:NewSpecialWarningMoveTo(360012, false, nil, nil, 1, 2)--Off by default because person has to actually have basic understanding of mechanic first, then agree to this helpful warning to help with it
 local yellCloudofCarrion						= mod:NewYell(360012)
-local yellBitingWounds							= mod:NewIconRepeatYell(364985, DBM_CORE_L.AUTO_YELL_ANNOUNCE_TEXT.shortyell)
+local yellBitingWounds							= mod:NewIconRepeatYell(364985)
 local specWarnLeechingClaws						= mod:NewSpecialWarningDefensive(359960, nil, nil, nil, 1, 2)
 local specWarnOpenedVeins						= mod:NewSpecialWarningTaunt(359963, nil, nil, nil, 1, 2)
 ----Shadow adds
@@ -71,7 +71,7 @@ mod:GroupSpells(360012, 364985)--Group biting wounds with cloud of carrion
 --Kin'tessa
 mod:AddTimerLine(DBM:EJ_GetSectionInfo(23929))
 local warnShatterMind							= mod:NewSpellAnnounce(360420, 4)--Kind of a generic alert to say "this pull is a wash"
-local warnFearfulTrepidation					= mod:NewTargetNoFilterAnnounce(360146, 3, nil, nil, 39176)
+local warnFearfulTrepidation					= mod:NewTargetCountAnnounce(360146, 3, nil, nil, 39176, nil, nil, nil, true)
 local warnAuraofShadows							= mod:NewSpellAnnounce(363191, 4)
 local warnAuraofShadowsOver						= mod:NewEndAnnounce(363191, 1)
 local warnSlumberCloud							= mod:NewCountAnnounce(360229, 2)
@@ -259,7 +259,7 @@ end
 
 function mod:SPELL_CAST_SUCCESS(args)
 	local spellId = args.spellId
-	if spellId == 360420 then
+	if spellId == 360420 and self:AntiSpam(3, 1) then
 		warnShatterMind:Show()
 	end
 end
@@ -329,7 +329,7 @@ function mod:SPELL_AURA_APPLIED(args)
 		end
 		timerFearfulTrepidationCD:Pause(self.vb.fearfulCount+1)
 	elseif spellId == 360418 and args:IsPlayer() then
-		timerParanoia:Start()
+		timerParanoia:Start(self:IsEasy() and 40 or 25)
 	elseif spellId == 360146 then
 		local icon = self.vb.trepidationIcon
 		if self.Options.SetIconOnFearfulTrepidation then
@@ -347,7 +347,7 @@ function mod:SPELL_AURA_APPLIED(args)
 			specWarnCloudofCarrionDebuffMove:CombinedShow(0.5, args.destName)
 			specWarnCloudofCarrionDebuffMove:ScheduleVoice(0.5, "gathershare")
 		end
-		warnFearfulTrepidation:CombinedShow(0.5, args.destName)
+		warnFearfulTrepidation:CombinedShow(0.5, self.vb.fearfulCount, args.destName)
 		self.vb.trepidationIcon = self.vb.trepidationIcon + 1
 	elseif spellId == 360148 then
 		if args:IsPlayer() then