+++ AddOns/ElvUI_BagFilter/CHANGELOG.md
@@ -1,6 +1,6 @@
 # ElvUI Bag Filters
 
-## [v7.4](https://github.com/warmexx/elvui-bag-filters/tree/v7.4) (2021-11-14)
-[Full Changelog](https://github.com/warmexx/elvui-bag-filters/commits/v7.4) [Previous Releases](https://github.com/warmexx/elvui-bag-filters/releases)
+## [v7.5](https://github.com/warmexx/elvui-bag-filters/tree/v7.5) (2022-04-03)
+[Full Changelog](https://github.com/warmexx/elvui-bag-filters/commits/v7.5) [Previous Releases](https://github.com/warmexx/elvui-bag-filters/releases)
 
-- Bumps TOC  
+- Updates TOC